+++ EldritchArcana/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.4]
+
+### Fixed
+- Update mod to work against game version 1.2.3 (other fixes planned for 0.9.4
+  will be released as 0.9.5, to get this out ASAP).
+
+
 ## [0.9.3]
 
 ### Fixed